Update catalogue of current files
Update the files Contents.m and somtoolbox.m
These files both contain the exact same data about the files and folders
in SOM-Toolbox. They were updated to reflect the new structure of v2.1

Additionally the floating file for rneural_gas was put into a folder.

Added some notes to som_make.m and som_show.m for problems with the
toolbox in Octave.